Add an input_filter for TextEdit.

+++ crates/egui/src/widgets/text_edit/builder.rs
@@ -58,6 +58,7 @@ pub struct TextEdit<'t> {
     font_selection: FontSelection,
     text_color: Option<Color32>,
     layouter: Option<&'t mut dyn FnMut(&Ui, &str, f32) -> Arc<Galley>>,
+    input_filter: Option<&'t mut dyn FnMut(&str) -> bool>,
     password: bool,
     frame: bool,
     margin: Vec2,
@@ -108,6 +109,7 @@ impl<'t> TextEdit<'t> {
             font_selection: Default::default(),
             text_color: None,
             layouter: None,
+            input_filter: None,
             password: false,
             frame: true,
             margin: vec2(4.0, 2.0),
@@ -222,6 +224,12 @@ impl<'t> TextEdit<'t> {
     /// ```
     pub fn layouter(mut self, layouter: &'t mut dyn FnMut(&Ui, &str, f32) -> Arc<Galley>) -> Self {
         self.layouter = Some(layouter);
+
+        self
+    }
+
+    pub fn input_filter(mut self, input_filter: &'t mut dyn FnMut(&str) -> bool) -> Self {
+        self.input_filter = Some(input_filter);
 
         self
     }
@@ -410,6 +418,7 @@ impl<'t> TextEdit<'t> {
             font_selection,
             text_color,
             layouter,
+            input_filter,
             password,
             frame: _,
             margin,
@@ -457,6 +466,10 @@ impl<'t> TextEdit<'t> {
         let layouter = layouter.unwrap_or(&mut default_layouter);
 
         let mut galley = layouter(ui, text.as_str(), wrap_width);
+        let mut default_input_filter = |_: &str| {
+            true
+        };
+        let input_filter = input_filter.unwrap_or(&mut default_input_filter);
 
         let desired_width = if clip_text {
             wrap_width // visual clipping with scroll in singleline input.
@@ -587,6 +600,7 @@ impl<'t> TextEdit<'t> {
                 text,
                 &mut galley,
                 layouter,
+                input_filter,
                 id,
                 wrap_width,
                 multiline,
@@ -878,6 +892,7 @@ fn events(
     text: &mut dyn TextBuffer,
     galley: &mut Arc<Galley>,
     layouter: &mut dyn FnMut(&Ui, &str, f32) -> Arc<Galley>,
+    input_filter: &mut dyn FnMut(&str) -> bool,
     id: Id,
     wrap_width: f32,
     multiline: bool,
@@ -923,7 +938,7 @@ fn events(
                 }
             }
             Event::Paste(text_to_insert) => {
-                if !text_to_insert.is_empty() {
+                if !text_to_insert.is_empty() && input_filter(text_to_insert) {
                     let mut ccursor = delete_selected(text, &cursor_range);
 
                     insert_text(&mut ccursor, text, text_to_insert, char_limit);
@@ -935,7 +950,7 @@ fn events(
             }
             Event::Text(text_to_insert) => {
                 // Newlines are handled by `Key::Enter`.
-                if !text_to_insert.is_empty() && text_to_insert != "\n" && text_to_insert != "\r" {
+                if !text_to_insert.is_empty() && text_to_insert != "\n" && text_to_insert != "\r" && input_filter(text_to_insert) {
                     let mut ccursor = delete_selected(text, &cursor_range);
 
                     insert_text(&mut ccursor, text, text_to_insert, char_limit);
@@ -951,7 +966,7 @@ fn events(
                 modifiers,
                 ..
             } => {
-                if multiline && ui.memory(|mem| mem.has_lock_focus(id)) {
+                if multiline && ui.memory(|mem| mem.has_lock_focus(id)) && input_filter("\t") {
                     let mut ccursor = delete_selected(text, &cursor_range);
                     if modifiers.shift {
                         // TODO(emilk): support removing indentation over a selection?
@@ -1013,7 +1028,7 @@ fn events(
             Event::CompositionUpdate(text_mark) => {
                 // empty prediction can be produced when user press backspace
                 // or escape during ime. We should clear current text.
-                if text_mark != "\n" && text_mark != "\r" && state.has_ime {
+                if text_mark != "\n" && text_mark != "\r" && state.has_ime && input_filter(text_mark) {
                     let mut ccursor = delete_selected(text, &cursor_range);
                     let start_cursor = ccursor;
                     if !text_mark.is_empty() {
@@ -1026,7 +1041,7 @@ fn events(
             }
 
             Event::CompositionEnd(prediction) => {
-                if prediction != "\n" && prediction != "\r" && state.has_ime {
+                if prediction != "\n" && prediction != "\r" && state.has_ime && input_filter(prediction) {
                     state.has_ime = false;
                     let mut ccursor = delete_selected(text, &cursor_range);
                     if !prediction.is_empty() {
